Add UID index length to healthcheck.
The UID index contains one UID per plone object. The UID for each object is
unique, thus we know that for each object in the catalog there will also be one
entry in the UID index. Thus we can use this index as a sample to check the
catalog index health.