Fix for acquired objects moved into their parents.
Correct special case handling for objects that were moved into one of their
parent objects and thus can still be found via acquisition when their former
path is used as their id was not changed.
We assumed that the redirector is involved, this is not correct. Thus drop
redirector check and instead make sure that the acquisition chain of the wrapped
and unwrapped object is different.