Use remove extra rid surgery for new symptoms.
Sometimes the UUID index is in a different state where the index
seems entirely stale. In such cases two different symptoms are
discovered, the extra RID (the old rid) and also a missing entry in the
UUID index (for the new rid). Removing the old rid and then reindexing
the UUID index for the affected objects seems to help in such cases.