Drop safeguard that compares direct parent.
This doesn't work when an object has been moved together with its
parent. The safeguard was initially implemented to add an extra level of
paranoia, we now loose it to only check if the new object path is
a shorter version of the old path. That way we know the object has been
moved up in the tree.

Maybe this restriction can be loosened a bit more eventually, but for
now it covers what we've seen in production.