Implement reindexing objects for orphaned rids.
There are cases where the rid is orphaned in the catalog but the object
is still around. We cleanup traces of the old object first, then reindex
it to create a new rid in the catalog.

We could solve this differently and attempt to repair the existing
record instead of creating a new one. But since the path->rid mapping
is the primary source of truth for the zcatalog i'd rather create a new
rid for such cases.