Drop specific surgery in favour of generic one.
Turns out the specific surgery was just a special case of the generic
one and can be removed/replaced with the generic surgery. This avoids
having to duplicate all special cases in two surgeries. Also cases
have surfaced where the object needs a full reindex in all indices, not
only the UID index for the same symptoms.